Move setPurchaseAutoRenew to upgrades data file
